requirements.txt
streamlit>=1.38.0
pandas>=2.2.0
yfinance>=0.2.43
altair>=5.4.1
pykrx>=1.0.51
python-dotenv
sqlalchemy
psycopg2-binary
